+++ src/main/java/LadderApplication.java
@@ -4,7 +4,7 @@ import repository.results.ResultsRepository;
 import repository.users.UserRepository;
 import service.ladder.LadderServiceImpl;
 import service.results.ResultsServiceImpl;
-import service.users.UserServiceImpl;
+import service.users.UserService;
 import view.InputView;
 import view.OutputView;
 
@@ -12,7 +12,7 @@ public class LadderApplication {
     public static void main(String[] args) {
         LadderController ladderController = new LadderController(
                 new LadderServiceImpl(new LadderRepository()),
-                new UserServiceImpl(new UserRepository()),
+                new UserService(new UserRepository()),
                 new ResultsServiceImpl(new ResultsRepository()),
                 new OutputView(),
                 new InputView()
